reorg topography
