+++ pulse2percept/topography/retina/__init__.py
@@ -1,0 +1,20 @@
+"""Retinal visual field maps
+
+.. autosummary::
+    :toctree: _api
+
+    base
+    curcio1990
+    watson2014
+
+"""
+from .base import RetinalMap
+from .curcio1990 import Curcio1990Map
+from .watson2014 import Watson2014Map, Watson2014DisplaceMap
+
+__all__ = [
+    'Curcio1990Map',
+    'RetinalMap',
+    'Watson2014DisplaceMap',
+    'Watson2014Map',
+]
