--- pulse2percept/topography/tests/test_retina.py
@@ -1,152 +1,0 @@
-import copy
-import numpy as np
-import pytest
-import numpy.testing as npt
-
-from pulse2percept.topography import (Curcio1990Map, Watson2014Map,
-                                 Watson2014DisplaceMap)
-from pulse2percept.units import (DimensionMismatchError, Quantity, dva, mm,
-                                 ms, um)
-
-
-def test_Curcio1990Map():
-    # Curcio1990 uses a linear dva_to_ret conversion factor:
-    for factor in [0.0, 1.0, 2.0]:
-        npt.assert_almost_equal(Curcio1990Map().dva_to_ret(factor, factor),
-                                (280.0 * factor, -280.0 * factor))
-    for factor in [0.0, 1.0, 2.0]:
-        npt.assert_almost_equal(Curcio1990Map().ret_to_dva(280.0 * factor,
-                                                      -280.0 * factor),
-                                (factor, factor))
-
-
-def test_eq_Curcio19990Map():
-    curcio_map = Curcio1990Map()
-
-    # Assert not equal for differing classes
-    npt.assert_equal(curcio_map == int, False)
-
-    # Assert equal to itself
-    npt.assert_equal(curcio_map == curcio_map, True)
-
-    # Assert equal for shallow references
-    copied = curcio_map
-    npt.assert_equal(curcio_map == copied, True)
-
-    # Assert deep copies are equal
-    copied = copy.deepcopy(curcio_map)
-    npt.assert_equal(curcio_map == copied, True)
-
-    # Assert differing objects aren't equal
-    differing_map = Watson2014Map()
-    npt.assert_equal(curcio_map == differing_map, False)
-
-
-def test_Watson2014Map():
-    trafo = Watson2014Map()
-    with pytest.raises(ValueError):
-        trafo.ret_to_dva(0, 0, coords='invalid')
-    with pytest.raises(ValueError):
-        trafo.dva_to_ret(0, 0, coords='invalid')
-
-    # Below 15mm eccentricity, relationship is linear with slope 3.731
-    npt.assert_almost_equal(trafo.ret_to_dva(0.0, 0.0), (0.0, 0.0))
-    for sign in [-1, 1]:
-        for exp in [2, 3, 4]:
-            ret = sign * 10 ** exp  # mm
-            dva = -3.731 * sign * 10 ** (exp - 3)  # dva
-            npt.assert_almost_equal(trafo.ret_to_dva(0, ret)[1], dva,
-                                    decimal=3 - exp)  # adjust precision
-    # Below 50deg eccentricity, relationship is linear with slope 0.268
-    npt.assert_almost_equal(trafo.dva_to_ret(0.0, 0.0), (0.0, 0.0))
-    for sign in [-1, 1]:
-        for exp in [-2, -1, 0]:
-            dva = sign * 10 ** exp  # deg
-            ret = -0.268 * sign * 10 ** (exp + 3)  # mm
-            npt.assert_almost_equal(trafo.dva_to_ret(0, dva)[1], ret,
-                                    decimal=-exp)  # adjust precision
-
-
-def test_eq_Watson2014Map():
-    map = Watson2014Map()
-
-    # Assert not equal for differing classes
-    npt.assert_equal(map == int, False)
-
-    # Assert equal to itself
-    npt.assert_equal(map == map, True)
-
-    # Assert equal for shallow references
-    copied = map
-    npt.assert_equal(map == copied, True)
-
-    # Assert deep copies are equal
-    copied = copy.deepcopy(map)
-    npt.assert_equal(map == copied, True)
-
-    # Assert differing objects aren't equal
-    differing_map = Curcio1990Map()
-    npt.assert_equal(map == differing_map, False)
-
-
-def test_Watson2014DisplaceMap():
-    trafo = Watson2014DisplaceMap()
-    with pytest.raises(ValueError):
-        trafo.watson_displacement(0, meridian='invalid')
-    npt.assert_almost_equal(trafo.watson_displacement(0), 0.4957506)
-    npt.assert_almost_equal(trafo.watson_displacement(100), 0)
-
-    # Check the max of the displacement function for the temporal meridian:
-    radii = np.linspace(0, 30, 100)
-    all_displace = trafo.watson_displacement(radii, meridian='temporal')
-    npt.assert_almost_equal(np.max(all_displace), 2.153532)
-    npt.assert_almost_equal(radii[np.argmax(all_displace)], 1.8181818)
-
-    # Check the max of the displacement function for the nasal meridian:
-    all_displace = trafo.watson_displacement(radii, meridian='nasal')
-    npt.assert_almost_equal(np.max(all_displace), 1.9228664)
-    npt.assert_almost_equal(radii[np.argmax(all_displace)], 2.1212121)
-    # Smoke test
-    trafo.dva_to_ret(0, 0)
-
-
-def test_retinal_map_units():
-    """dva in, microns out -- and either may be spelled with a unit"""
-    for cls in (Curcio1990Map, Watson2014Map, Watson2014DisplaceMap):
-        visual_field_map = cls()
-        bare = visual_field_map.dva_to_ret(5, -2)
-        npt.assert_allclose(visual_field_map.dva_to_ret(5 * dva, -2 * dva),
-                            bare,
-                            rtol=1e-12, err_msg=cls.__name__)
-        # The output is a plain number of microns, never a Quantity:
-        for value in bare:
-            npt.assert_equal(isinstance(value, Quantity), False)
-        with pytest.raises(DimensionMismatchError):
-            visual_field_map.dva_to_ret(5 * um, -2)
-        with pytest.raises(DimensionMismatchError):
-            visual_field_map.dva_to_ret(5, -2 * ms)
-
-    # The inverse takes a length, and mixed spellings round-trip:
-    for cls in (Curcio1990Map, Watson2014Map):
-        visual_field_map = cls()
-        x_um, y_um = visual_field_map.dva_to_ret(5 * dva, -2 * dva)
-        bare = visual_field_map.ret_to_dva(x_um, y_um)
-        mixed = visual_field_map.ret_to_dva((x_um / 1000) * mm, y_um * um)
-        npt.assert_allclose(mixed, bare, rtol=1e-9, err_msg=cls.__name__)
-        with pytest.raises(DimensionMismatchError):
-            visual_field_map.ret_to_dva(5 * dva, -2)
-    # Curcio is exactly linear, so its round trip closes exactly. (Watson's
-    # forward and inverse are separate fits and only agree to ~2%, which is a
-    # property of that map and not of the unit conversion.)
-    npt.assert_allclose(
-        Curcio1990Map().ret_to_dva(*Curcio1990Map().dva_to_ret(5 * dva,
-                                                               -2 * dva)),
-        [5, -2], rtol=1e-12)
-
-    # A non-coordinate keyword travels through untouched:
-    watson = Watson2014Map()
-    npt.assert_allclose(watson.dva_to_ret(3 * dva, 1 * dva, coords='cart'),
-                        watson.dva_to_ret(3, 1, coords='cart'), rtol=1e-12)
-    npt.assert_allclose(watson.ret_to_dva(1000 * um, 500 * um, coords='cart'),
-                        watson.ret_to_dva(1000, 500, coords='cart'),
-                        rtol=1e-12)
